refactor: update home pages to new structure and remove deprecated files
