+++ app/(homes)/home-1/page.jsx
@@ -1,37 +1,44 @@
-import Header1 from "@/components/headers/Header1";
-import Features from "@/components/homes/home-1/Features";
-import Hero from "@/components/homes/home-1/Hero";
-import Integretion from "@/components/homes/home-1/Integretion";
-import Facts from "@/components/homes/home-1/Facts";
-import Pricing from "@/components/homes/home-1/Pricing";
-import Feedback from "@/components/homes/home-1/Feedback";
-import Timeline from "@/components/homes/home-1/Timeline";
-import Team from "@/components/homes/home-1/Team";
-import Cta from "@/components/homes/home-1/Cta";
-import Footer1 from "@/components/footers/Footer1";
+import Footer7 from "@/components/footers/Footer7";
+import Header7 from "@/components/headers/Header7";
+import Blogs from "@/components/homes/home-8/Blogs";
+import Brands from "@/components/homes/home-8/Brands";
+import Cta from "@/components/homes/home-8/Cta";
+import Faqs from "@/components/homes/home-8/Faqs";
+import Features from "@/components/homes/home-8/Features";
+import Features2 from "@/components/homes/home-8/Features2";
+import Hero from "@/components/homes/home-8/Hero";
+import Pricing from "@/components/homes/home-8/Pricing";
+import Testimonials from "@/components/homes/home-8/Testimonials";
+import Tools from "@/components/homes/home-8/Tools";
+import React from "react";
 export const metadata = {
   title:
-    "Home 1 || Sendbix - Full-featured, professional-looking software, saas and startup nextjs template.",
+    "Home 8 || Sendbix - Full-featured, professional-looking software, saas and startup nextjs template.",
   description:
     "Sendbix - Full-featured, professional-looking software, saas and startup nextjs template.",
 };
-export default function HomePage1() {
+export default function page() {
   return (
     <>
-      <div className="uni-body panel bg-white text-gray-900 dark:bg-gray-900 dark:text-gray-200 overflow-x-hidden bp-xs bp-sm bp-md bp-lg bp-xl bp-xxl-max dom-ready">
-        <Header1 />
-        <div id="wrapper" className="wrap">
-          <Hero />
-          <Features />
-          <Integretion />
-          <Facts />
-          <Pricing />
-          <Feedback />
-          <Timeline />
-          <Team />
-          <Cta />
+      <div className="theme-4">
+        <div className="page-wrapper">
+          <div className="bp-xs bp-sm bp-md bp-lg dom-ready bp-xxl-max bp-xl uni-body panel uni-body panel bg-white text-tertiary-900 dark:bg-tertiary-800 dark:text-tertiary-200 overflow-x-hidden disable-cursor">
+            <Header7 />
+            <div id="wrapper" className="wrap">
+              <Hero />
+              <Brands />
+              <Features />
+              <Tools />
+              <Features2 />
+              <Pricing />
+              <Testimonials />
+              <Faqs />
+              <Blogs />
+              <Cta />
+            </div>
+            <Footer7 />
+          </div>{" "}
         </div>
-        <Footer1 />
       </div>
     </>
   );
